TiltNetworkTool-1.0.0-macOSfix/requirements.txt
PyQt6
PyOpenGL
numpy
pyrr
pyserial
pyqtgraph